sotp: Add 1.0.0-unstable-2024-05-20
Version 1.0.0 doesn't build because of a dependency problem.

```
missing go.sum entry for module providing package
google.golang.org/appengine/urlfetch (imported by golang.org/x/oauth2/internal)
```

The upstream repository has an open PR to license the project under
MPL-2.0 but it's on hold because they might opt for Apache-2.0 instead.
The project is clearly intended to be licensed under an open license but
until getsops/sotp#1 is resolved it is to be considered unfree software.